djangoBasic/requirement.txt
decorator==5.1.1
Django==4.2.17
djangorestframework==3.15.2
ipdb==0.13.13
ipython==8.12.3